backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic-settings==2.5.2
pandas==2.2.3
numpy==2.1.1
MetaTrader5==5.0.4424
pytest==8.3.3
httpx==0.27.2
